+++ src/radio.c
@@ -8,7 +8,7 @@
  * @copyright Copyright (c) 2024
  * 
  */
-#include "radio_protocol.h"
+#include "radio.h"
 
 #include "stdio.h"
